requirements.txt
numpy
scipy
pandas
matplotlib
scikit-learn
torch
torchvision
tqdm
requests
pyyaml
networkx
arxiv
bibtexparser
markdown
pyppeteer
gradio==3.50.2
huggingface_hub==0.22.2
fpdf2
